0. Что такое тестирование [#159431]

+++ chapter_001/src/test/java/ru/job4j/converter/ConverterTest.java
@@ -9,28 +9,28 @@ public class ConverterTest {
     @Test
     public void when60RubleToDollarThen1() {
         Converter converter = new Converter();
-        int result = converter.rubleToDollar(60);
+        float result = converter.rubleToDollar(60);
         assertThat(result, is(1));
     }
 
     @Test
     public void when70RubleToEuroThen1() {
         Converter converter = new Converter();
-        int result = converter.rubleToEuro(70);
+        float result = converter.rubleToEuro(70);
         assertThat(result, is(1));
     }
 
     @Test
     public void when1DollarToRubleThen60() {
         Converter converter = new Converter();
-        int result = converter.dollarToRuble(1);
+        float result = converter.dollarToRuble(1);
         assertThat(result, is(60));
     }
 
     @Test
     public void when1EuroToRubleThen70() {
         Converter converter = new Converter();
-        int result = converter.euroToRuble(1);
+        float result = converter.euroToRuble(1);
         assertThat(result, is(70));
     }
 }
